@@ -1,15 +1,5 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using API.Data;
-using API.Models;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
-
-namespace API.Controllers
-{
-    public class BuggyController : BaseApiController
+namespace API.Controllers;
+public class BuggyController : BaseApiController
     {
         private readonly DataContext _context;
         public BuggyController(DataContext context){
@@ -47,4 +37,3 @@ namespace API.Controllers
             return BadRequest();
         }
     }
-}
