@@ -1,62 +1,49 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using API.DTOs;
-using API.Extensions;
-using API.Helpers;
-using API.Interfaces;
-using API.Models;
-using Microsoft.AspNetCore.Mvc;
-
-namespace API.Controllers
+namespace API.Controllers;
+[ApiController]
+[Route("api/[controller]")]
+public class LikesController : BaseApiController
 {
-    [ApiController]
-    [Route("api/[controller]")]
-    public class LikesController : BaseApiController
+    private readonly IUnitOfWork _unitOfWork;
+
+    public LikesController(IUnitOfWork unitOfWork)
     {
-        private readonly IUnitOfWork _unitOfWork;
+        _unitOfWork = unitOfWork;
+    }
+    [HttpPost("{username}")]
+    public async Task<ActionResult> AddLike(string username)
+    {
+        var sourceUserId = User.GetUserId();
+        var likedUser = await _unitOfWork.UserRepository.GetUserByUsernameAsync(username);
+        var sourceUser = await _unitOfWork.LikesRepository.GetUserWithLikes(sourceUserId);
 
-        public LikesController(IUnitOfWork unitOfWork)
+        if(likedUser == null) return NotFound();
+
+        if(sourceUser.UserName == username) return BadRequest("You cannot like yourself");
+
+        var userLike = await _unitOfWork.LikesRepository.GetUserLike(sourceUserId, likedUser.Id);
+
+        if(userLike != null) return BadRequest("You already like this user");
+
+        userLike = new UserLike
         {
-            _unitOfWork = unitOfWork;
-        }
-        [HttpPost("{username}")]
-        public async Task<ActionResult> AddLike(string username)
-        {
-            var sourceUserId = User.GetUserId();
-            var likedUser = await _unitOfWork.UserRepository.GetUserByUsernameAsync(username);
-            var sourceUser = await _unitOfWork.LikesRepository.GetUserWithLikes(sourceUserId);
+            SourceUserId = sourceUserId,
+            LikedUserId = likedUser.Id
+        };
+        sourceUser.LikedUsers.Add(userLike);
 
-            if(likedUser == null) return NotFound();
+        if(await _unitOfWork.Complete()) return Ok();
 
-            if(sourceUser.UserName == username) return BadRequest("You cannot like yourself");
+        return BadRequest("Failed to like user");
+    }
 
-            var userLike = await _unitOfWork.LikesRepository.GetUserLike(sourceUserId, likedUser.Id);
+    [HttpGet]
+    public async Task<ActionResult<IEnumerable<LikeDto>>> GetUserLikes([FromQuery]LikesParams likesParams)
+    {
+        likesParams.UserId = User.GetUserId();
+        var users = await _unitOfWork.LikesRepository.GetUserLikes(likesParams);
 
-            if(userLike != null) return BadRequest("You already like this user");
+        Response.AddPaginationHeader(users.CurrentPage, users.PageSize, users.TotalCount,users.TotalPages);
 
-            userLike = new UserLike
-            {
-                SourceUserId = sourceUserId,
-                LikedUserId = likedUser.Id
-            };
-            sourceUser.LikedUsers.Add(userLike);
-
-            if(await _unitOfWork.Complete()) return Ok();
-
-            return BadRequest("Failed to like user");
-        }
-
-        [HttpGet]
-        public async Task<ActionResult<IEnumerable<LikeDto>>> GetUserLikes([FromQuery]LikesParams likesParams)
-        {
-            likesParams.UserId = User.GetUserId();
-            var users = await _unitOfWork.LikesRepository.GetUserLikes(likesParams);
-
-            Response.AddPaginationHeader(users.CurrentPage, users.PageSize, users.TotalCount,users.TotalPages);
-
-            return Ok(users);
-        }
+        return Ok(users);
     }
 }
