@@ -1,60 +1,45 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using API.DTOs;
-using API.Extensions;
-using API.Helpers;
-using API.Interfaces;
-using API.Models;
-using AutoMapper;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
-
-namespace API.Controllers
+namespace API.Controllers;
+[Authorize]
+public class MessagesController : BaseApiController
 {
-    [Authorize]
-    public class MessagesController : BaseApiController
+    private readonly IUnitOfWork _unitOfWork;
+    private readonly IMapper _mapper;
+    public MessagesController(IUnitOfWork unitOfWork , IMapper mapper)
     {
-        private readonly IUnitOfWork _unitOfWork;
-        private readonly IMapper _mapper;
-        public MessagesController(IUnitOfWork unitOfWork , IMapper mapper)
-        {
-            _unitOfWork = unitOfWork;
-            _mapper = mapper;
-        }
+        _unitOfWork = unitOfWork;
+        _mapper = mapper;
+    }
 
-        [HttpGet]
-        public async Task<ActionResult<IEnumerable<MessageDto>>> GetMessageForUser([FromQuery] MessageParams messageParams)
-        {
-            messageParams.Username = User.GetUsername();
+    [HttpGet]
+    public async Task<ActionResult<IEnumerable<MessageDto>>> GetMessageForUser([FromQuery] MessageParams messageParams)
+    {
+        messageParams.Username = User.GetUsername();
 
-            var messages = await _unitOfWork.MessageRepository.GetMessageForUser(messageParams);
+        var messages = await _unitOfWork.MessageRepository.GetMessageForUser(messageParams);
 
-            Response.AddPaginationHeader(messages.CurrentPage, messages.PageSize, 
-                messages.TotalCount, messages.TotalPages);
+        Response.AddPaginationHeader(messages.CurrentPage, messages.PageSize, 
+            messages.TotalCount, messages.TotalPages);
 
-            return messages;
-        }
+        return messages;
+    }
 
-        [HttpDelete("{id}")]
-        public async Task<ActionResult> DeleteMessage(int id)
-        {
-            var username = User.GetUsername();
+    [HttpDelete("{id}")]
+    public async Task<ActionResult> DeleteMessage(int id)
+    {
+        var username = User.GetUsername();
 
-            var message = await _unitOfWork.MessageRepository.GetMessage(id);
+        var message = await _unitOfWork.MessageRepository.GetMessage(id);
 
-            if(message.Sender.UserName != username && message.Recipient.UserName != username) return Unauthorized();
+        if(message.Sender.UserName != username && message.Recipient.UserName != username) return Unauthorized();
 
-            if(message.Sender.UserName == username) message.SenderDeleted = true;
+        if(message.Sender.UserName == username) message.SenderDeleted = true;
 
-            if(message.Recipient.UserName == username) message.RecipientDeleted = true;
+        if(message.Recipient.UserName == username) message.RecipientDeleted = true;
 
-            if(message.SenderDeleted && message.RecipientDeleted) _unitOfWork.MessageRepository.DeleteMessage(message);
+        if(message.SenderDeleted && message.RecipientDeleted) _unitOfWork.MessageRepository.DeleteMessage(message);
 
-            if(await _unitOfWork.Complete()) return Ok();
+        if(await _unitOfWork.Complete()) return Ok();
 
-            return BadRequest("Problem deleting the message");
-        }
+        return BadRequest("Problem deleting the message");
     }
 }
