@@ -1,15 +1,4 @@
-﻿using API.DTOs;
-using API.Extensions;
-using API.Helpers;
-using API.Interfaces;
-using API.Models;
-using Microsoft.EntityFrameworkCore;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace API.Data
-{
+﻿namespace API.Data;
     public class LikesRepository : ILikesRepository
     {
         private readonly DataContext _context;
@@ -61,4 +50,4 @@ namespace API.Data
                 .FirstOrDefaultAsync(x => x.Id == userId);
         }
     }
-}
+
