@@ -1,17 +1,6 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using API.DTOs;
-using API.Helpers;
-using API.Interfaces;
-using API.Models;
-using AutoMapper;
-using AutoMapper.QueryableExtensions;
-using Microsoft.EntityFrameworkCore;
 
-namespace API.Data
-{
+
+namespace API.Data;
     public class MessageRepository : IMessageRepository
     {
         private readonly DataContext _context;
@@ -109,4 +98,3 @@ namespace API.Data
                 .FirstOrDefaultAsync();
         }
     }
-}
