@@ -1,17 +1,4 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using API.DTOs;
-using API.Helpers;
-using API.Interfaces;
-using API.Models;
-using AutoMapper;
-using AutoMapper.QueryableExtensions;
-using Microsoft.EntityFrameworkCore;
-
-namespace API.Data
-{
+namespace API.Data;
     public class UserRepository : IUserRepository
     {
         private readonly DataContext _context;
@@ -82,4 +69,3 @@ namespace API.Data
             _context.Entry(user).State = EntityState.Modified;
         }
     }
-}
