@@ -1,13 +1,9 @@
-using API.Helpers;
-using Microsoft.AspNetCore.Mvc;
+namespace API.Controllers;
 
-namespace API.Controllers
+[ServiceFilter(typeof(LogUserActivity))]
+[ApiController]
+[Route("api/[controller]")]
+public class BaseApiController : ControllerBase
 {
-    [ServiceFilter(typeof(LogUserActivity))]
-    [ApiController]
-    [Route("api/[controller]")]
-    public class BaseApiController : ControllerBase
-    {
         
-    }
 }
