@@ -1,10 +1,5 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
-using API.Helpers;
-using System.Text.Json;
+
+using JsonSerializer = System.Text.Json.JsonSerializer;
 
 namespace API.Extensions
 {
