@@ -1,10 +1,4 @@
-﻿using API.DTOs;
-using API.Helpers;
-using API.Models;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace API.Interfaces
+﻿namespace API.Interfaces
 {
     public interface ILikesRepository
     {
