@@ -1,6 +1,4 @@
-﻿using System.Threading.Tasks;
-
-namespace API.Interfaces
+﻿namespace API.Interfaces
 {
     public interface IUnitOfWork
     {
