@@ -1,8 +1,4 @@
-﻿using API.Extensions;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.SignalR;
-using System;
-using System.Threading.Tasks;
+﻿
 
 namespace API.SignalR
 {
