@@ -1,8 +1,4 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace API.SignalR
+﻿namespace API.SignalR
 {
     public class PresenceTracker
     {
